+++ views/checkout.php
@@ -14,14 +14,20 @@
 <html <?php language_attributes(); ?>>
 	<head>
 		<meta charset="<?php bloginfo( 'charset' ); ?>" />
+		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
 
 		<title><?php esc_html_e( 'Checkout', 'pronamic_ideal' ); ?></title>
 
 		<?php
 
-		wp_print_styles( 'pronamic-pay-redirect' );
-
-		wp_print_scripts( 'pronamic-pay-adyen-checkout' );
+		/**
+		 * Triggers on Adyen checkout page.
+		 *
+		 * @link https://github.com/wp-pay-gateways/adyen#pronamic_pay_adyen_checkout_head
+		 *
+		 * @since 1.1
+		 */
+		do_action( 'pronamic_pay_adyen_checkout_head' );
 
 		?>
 	</head>