+++ CHANGELOG.md
@@ -7,12 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased][unreleased]
 -
 
-## 1.0.1 - 2019-05-14
+## [1.0.2] - 2019-08-27
+- Set country from billing address.
+- Added action `pronamic_pay_adyen_checkout_head`.
+- Added `pronamic_pay_adyen_config_object` filter and improved documentation.
+
+## [1.0.1] - 2019-05-14
 - Remove path from origin URL in payment session request.
 - Fix API live URL prefix setting not saved.
 
 ## 1.0.0 - 2019-03-28
 - First release.
 
-[unreleased]: https://github.com/wp-pay-gateways/adyen/compare/1.0.1...HEAD
+[unreleased]: https://github.com/wp-pay-gateways/adyen/compare/1.0.2...HEAD
+[1.0.2]: https://github.com/wp-pay-gateways/adyen/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/wp-pay-gateways/adyen/compare/1.0.0...1.0.1